+++ src/actions/aiQuestion.ts
@@ -1,6 +1,6 @@
 import chalk from "chalk";
-import {saveResultForBashWrapper, saveUserInputInHistory} from "../util/fileUtil";
-import {askOpenAi} from "../util/aiUtil";
+import {saveResultForBashWrapper, saveUserInputInHistory} from "../services/fileManagement";
+import {askOpenAi} from "../services/openAi";
 
 export const askAi = (question: string) => {
     saveUserInputInHistory(question)
