+++ src/services/fileManagement.ts
@@ -1,14 +1,12 @@
 import path from 'path';
-import { atob, btoa } from 'buffer';
 import * as fs from 'fs';
 import chalk from 'chalk';
-import {prepareTextForSave} from "./historyUtil";
+import {prepareTextForSave} from "./history";
 
-export const getHome = () => path.resolve(process.env['HOME'] || process.env['USERPROFILE']);
-export const getHistoryPath = () => getHome() + '/.ai-history.txt';
-
+export const getHistoryPath = () => getHomeDir() + '/.ai-history.txt';
 
 export const saveResultForBashWrapper = (result: string) => fs.writeFileSync('/tmp/ai-autocomplete.out', result);
+
 
 export const clearHistory = (): void => {
   fs.truncateSync(getHistoryPath())
@@ -19,3 +17,5 @@ export const saveUserInputInHistory = (userInput: string) => {
   const text = prepareTextForSave(userInput)
   fs.appendFileSync(getHistoryPath(), `H: ${text}\n`);
 }
+
+export const getHomeDir = () => path.resolve(process.env['HOME'] || process.env['USERPROFILE']);
