--- src/util/tokenUtil.ts
@@ -1,3 +1,0 @@
-import {encode} from 'gpt-3-encoder';
-
-export const countTokens = (question: string) => encode(question).length
